Add files via upload
The project is a part of the Coursera IBM Applied AI certification. This project uses IBM Watson and Open CV to classify  different types of furniture. Watson was trained to identify each category (30 images each) of furniture with a confidence score of 75% and above. Watson was then tested with new images (previously unused for training), to confirm that it was able to differentiate between these three types of furniture with a high confidence score. The results of this project have been made available here.